host-zephyr: Expose common functions for host DMA new/free
Create and expose a couple of functions for host DMA allocation and
free along with struct host_data,these functions will be shared by both
host and copier component.
This is in preparation to remove the creation of host component and
endpoint buffer in create_host().

Below is the preparation work:
1. Add host data structure to copier data structure
Currently, copier call host interface need through ops driver.
However, with dedicated work, all host interface can be called
through host data, this commit add host data into copier data.

2. Add Header file to share host data between copier and host
Data and interaface will be shared through this header file between
host and copier.
